chore: separated charts logic
